sna/video: Track xv buffers separately for each pipe
Since each Xv port can have a sprite on each crtc, and each
crtc can be individually rotated we need to start tracking the buffers
used by each sprite separately. That is if one crtc is rotate 90 or 270
degrees and another one is rotated 0 or 180 degress, and the sprites
can't natively do 90/270 the framebuffers for each sprite have to have
different dimensions.

To that end introduce an array of sna_video_crtc structures for each
Xv port and pull the relevant information down from the current per-port
structure.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>