sna: Disable DPMS-deferred modesetting when run as a secondary screen
When being run as a secondary screen, we do not receive DPMS events
(xf86DPMS() only operates on the primary screen, I guess one day we will
have to fix that and undo this hack). As such we cannot defer the
modesets until the final DPMS as it will never be sent.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=103618
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>