sna: Drop redundant sync of an inactive snoop cache
Having requested an inactive snooped buffer, we know that it is
already flushed and do not need to flush it again. With debugging
enabled, we hit an assert while flushing that the buffer has a refcout,
which at this present time of being plucked from the snoop cache it does
not.

Reported-by: Adric Blake <promarbler14@gmail.com>
References: https://bugs.freedesktop.org/show_bug.cgi?id=104341#c2
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>