sna/present: Only add the first element to the Timer queue
TimerSet scales linearly with the number of elements in the queue as it
performs an insertion sort, duplicating the sorting we also perform to
keep the per-crtc vblank queue in an orderly fashion. As we already
maintain the ordered timeline of vblanks, we can simply queue the next
when the current vblank completes.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>