sna/video/sprite: Remove the XV_ALWAYS_ON_TOP restriction for SKL+ scaling
On SKL+ the dst colorkey is enabled on the primary plane instead of the
sprite plane. That means the restriction of scaling vs. keying doesn't
actually apply here as we never scale the primary. So let's remove
the requirement of having XV_ALWAYS_ON_TOP enabled to get hw scaling.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>