sna/dri2: Interpret DRI2ATTACH_FORMAT as depth not bpp
In mesa i915/i965 pass the bpp to use when creating the surface, but the
gallium state tracker passed the depth. As it happens that
BitsPerPixel(format) will do the right thing for both, use that.

| DRI2ATTACH_FORMAT { attachment: CARD32
|		      format:     CARD32 }
|
|  The DRI2ATTACH_FORMAT describes an attachment and the associated
|  format.  'attachment' describes the attachment point for the buffer,
|  'format' describes an opaque, device-dependent format for the buffer.

Should we need to use an explicit format (heavens forbid as nobody likes
DRI2) then that will have to start in the range above 256 (or higher).

For now the convention is defined by the mixture of i965/iris, and that
is to assume it is essentially a depth.

Reported-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
References: https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/4569
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>